Revert "Merge branch 'main' into fix-readme-testnet-link"
This reverts commit 4634481cec38ddfb3eb61aaab97cc0dba01db903, reversing
changes made to 6365ba3491238bdc2f7589bd83a41e71e47f39f5.
