+++ .circleci/config.yml
@@ -831,7 +831,7 @@ jobs:
       - run:
           name: Build development contracts
           command: |
-            docker run --volumes-from with_code cosmwasm/workspace-optimizer:0.12.4
+            docker run --volumes-from with_code cosmwasm/workspace-optimizer:0.12.3
             docker cp with_code:/code/artifacts ./artifacts
       - run:
           name: Show data
@@ -848,7 +848,7 @@ jobs:
               -u "$CIRCLE_PROJECT_USERNAME" -r "$CIRCLE_PROJECT_REPONAME" \
               -c "$CIRCLE_SHA1" \
               -n "$TITLE" -b "$BODY" \
-              -replace \
+              -delete \
               "$TAG" ./artifacts/
 
   build_and_upload_schemas:
