+++ contracts/cw1-subkeys/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cw1-subkeys"
-version = "0.11.0"
+version = "0.10.3"
 authors = ["Ethan Frey <ethanfrey@users.noreply.github.com>"]
 edition = "2018"
 description = "Implement subkeys for authorizing native tokens as a cw1 proxy contract"
@@ -19,12 +19,12 @@ library = []
 test-utils = []
 
 [dependencies]
-utils = { path = "../../packages/utils", version = "0.11.0" }
-cw1 = { path = "../../packages/cw1", version = "0.11.0" }
-cw2 = { path = "../../packages/cw2", version = "0.11.0" }
-cw1-whitelist = { path = "../cw1-whitelist", version = "0.11.0", features = ["library"] }
+utils = { path = "../../packages/utils", version = "0.10.3" }
+cw1 = { path = "../../packages/cw1", version = "0.10.3" }
+cw2 = { path = "../../packages/cw2", version = "0.10.3" }
+cw1-whitelist = { path = "../cw1-whitelist", version = "0.10.3", features = ["library"] }
 cosmwasm-std = { version = "1.0.0-beta3", features = ["staking"] }
-cw-storage-plus = { path = "../../packages/storage-plus", version = "0.11.0" }
+cw-storage-plus = { path = "../../packages/storage-plus", version = "0.10.3" }
 schemars = "0.8.1"
 serde = { version = "1.0.103", default-features = false, features = ["derive"] }
 thiserror = "1.0.23"
@@ -32,4 +32,4 @@ semver = "1"
 
 [dev-dependencies]
 cosmwasm-schema = { version = "1.0.0-beta3" }
-cw1-whitelist = { path = "../cw1-whitelist", version = "0.11.0", features = ["library", "test-utils"] }
+cw1-whitelist = { path = "../cw1-whitelist", version = "0.10.3", features = ["library", "test-utils"] }
