+++ contracts/cw1-whitelist/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cw1-whitelist"
-version = "0.11.0"
+version = "0.10.3"
 authors = ["Ethan Frey <ethanfrey@users.noreply.github.com>"]
 edition = "2018"
 description = "Implementation of an proxy contract using a whitelist"
@@ -19,11 +19,11 @@ library = []
 test-utils = []
 
 [dependencies]
-utils = { path = "../../packages/utils", version = "0.11.0" }
-cw1 = { path = "../../packages/cw1", version = "0.11.0" }
-cw2 = { path = "../../packages/cw2", version = "0.11.0" }
+utils = { path = "../../packages/utils", version = "0.10.3" }
+cw1 = { path = "../../packages/cw1", version = "0.10.3" }
+cw2 = { path = "../../packages/cw2", version = "0.10.3" }
 cosmwasm-std = { version = "1.0.0-beta3", features = ["staking"] }
-cw-storage-plus = { path = "../../packages/storage-plus", version = "0.11.0" }
+cw-storage-plus = { path = "../../packages/storage-plus", version = "0.10.3" }
 schemars = "0.8.1"
 serde = { version = "1.0.103", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0.23" }
@@ -32,5 +32,5 @@ thiserror = { version = "1.0.23" }
 anyhow = "1"
 assert_matches = "1"
 cosmwasm-schema = { version = "1.0.0-beta3" }
-cw-multi-test = { path = "../../packages/multi-test", version = "0.11.0" }
+cw-multi-test = { path = "../../packages/multi-test", version = "0.10.3" }
 derivative = "2"
