+++ contracts/cw20-escrow/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cw20-escrow"
-version = "0.11.0"
+version = "0.10.3"
 authors = ["Ethan Frey <ethanfrey@users.noreply.github.com>"]
 edition = "2018"
 description = "Implementation of an escrow that accepts CosmWasm-20 tokens as well as native tokens"
@@ -18,16 +18,16 @@ backtraces = ["cosmwasm-std/backtraces"]
 library = []
 
 [dependencies]
-utils = { path = "../../packages/utils", version = "0.11.0" }
-cw2 = { path = "../../packages/cw2", version = "0.11.0" }
-cw20 = { path = "../../packages/cw20", version = "0.11.0" }
+utils = { path = "../../packages/utils", version = "0.10.3" }
+cw2 = { path = "../../packages/cw2", version = "0.10.3" }
+cw20 = { path = "../../packages/cw20", version = "0.10.3" }
 cosmwasm-std = { version = "1.0.0-beta3" }
-cw-storage-plus = { path = "../../packages/storage-plus", version = "0.11.0" }
+cw-storage-plus = { path = "../../packages/storage-plus", version = "0.10.3" }
 schemars = "0.8.1"
 serde = { version = "1.0.103", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0.23" }
 
 [dev-dependencies]
 cosmwasm-schema = { version = "1.0.0-beta3" }
-cw-multi-test = { path = "../../packages/multi-test", version = "0.11.0" }
-cw20-base = { path = "../cw20-base", version = "0.11.0", features = ["library"] }
+cw-multi-test = { path = "../../packages/multi-test", version = "0.10.3" }
+cw20-base = { path = "../cw20-base", version = "0.10.3", features = ["library"] }
