+++ contracts/cw20-ics20/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cw20-ics20"
-version = "0.11.0"
+version = "0.10.3"
 authors = ["Ethan Frey <ethanfrey@users.noreply.github.com>"]
 edition = "2018"
 description = "IBC Enabled contracts that receives CW20 tokens and sends them over ICS20 to a remote chain"
@@ -18,11 +18,11 @@ backtraces = ["cosmwasm-std/backtraces"]
 library = []
 
 [dependencies]
-utils = { path = "../../packages/utils", version = "0.11.0" }
-cw2 = { path = "../../packages/cw2", version = "0.11.0" }
-cw20 = { path = "../../packages/cw20", version = "0.11.0" }
+utils = { path = "../../packages/utils", version = "0.10.3" }
+cw2 = { path = "../../packages/cw2", version = "0.10.3" }
+cw20 = { path = "../../packages/cw20", version = "0.10.3" }
 cosmwasm-std = { version = "1.0.0-beta3", features = ["stargate"] }
-cw-storage-plus = { path = "../../packages/storage-plus", version = "0.11.0" }
+cw-storage-plus = { path = "../../packages/storage-plus", version = "0.10.3" }
 schemars = "0.8.1"
 serde = { version = "1.0.103", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0.23" }
