+++ contracts/cw20-staking/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cw20-staking"
-version = "0.11.0"
+version = "0.10.3"
 authors = ["Ethan Frey <ethanfrey@users.noreply.github.com>"]
 edition = "2018"
 description = "Implement simple staking derivatives as a cw20 token"
@@ -20,13 +20,13 @@ backtraces = ["cosmwasm-std/backtraces"]
 library = []
 
 [dependencies]
-utils = { path = "../../packages/utils", version = "0.11.0" }
-cw2 = { path = "../../packages/cw2", version = "0.11.0" }
-cw20 = { path = "../../packages/cw20", version = "0.11.0" }
-cw-controllers = { path = "../../packages/controllers", version = "0.11.0" }
-cw20-base = { path = "../../contracts/cw20-base", version = "0.11.0", features = ["library"] }
+utils = { path = "../../packages/utils", version = "0.10.3" }
+cw2 = { path = "../../packages/cw2", version = "0.10.3" }
+cw20 = { path = "../../packages/cw20", version = "0.10.3" }
+cw-controllers = { path = "../../packages/controllers", version = "0.10.3" }
+cw20-base = { path = "../../contracts/cw20-base", version = "0.10.3", features = ["library"] }
 cosmwasm-std = { version = "1.0.0-beta3", features = ["staking"] }
-cw-storage-plus = { path = "../../packages/storage-plus", version = "0.11.0" }
+cw-storage-plus = { path = "../../packages/storage-plus", version = "0.10.3" }
 schemars = "0.8.1"
 serde = { version = "1.0.103", default-features = false, features = ["derive"] }
 thiserror = { version = "1.0.23" }
