+++ contracts/cw3-fixed-multisig/src/error.rs
@@ -1,15 +1,10 @@
 use cosmwasm_std::StdError;
-use utils::ThresholdError;
-
 use thiserror::Error;
 
 #[derive(Error, Debug, PartialEq)]
 pub enum ContractError {
     #[error("{0}")]
     Std(#[from] StdError),
-
-    #[error("{0}")]
-    Threshold(#[from] ThresholdError),
 
     #[error("Required weight cannot be zero")]
     ZeroWeight {},
