+++ contracts/cw3-fixed-multisig/src/integration_tests.rs
@@ -1,14 +1,13 @@
 #![cfg(test)]
 
+use crate::contract::{execute, instantiate, query};
+use crate::msg::{ExecuteMsg, InstantiateMsg, Voter};
 use cosmwasm_std::{to_binary, Addr, Empty, Uint128, WasmMsg};
 use cw20::{BalanceResponse, MinterResponse};
 use cw20_base::msg::QueryMsg;
 use cw3::Vote;
 use cw_multi_test::{App, Contract, ContractWrapper, Executor};
-use utils::{Duration, Threshold};
-
-use crate::contract::{execute, instantiate, query};
-use crate::msg::{ExecuteMsg, InstantiateMsg, Voter};
+use utils::Duration;
 
 fn mock_app() -> App {
     App::default()
@@ -54,7 +53,7 @@ fn cw3_controls_cw20() {
                 weight: 1,
             },
         ],
-        threshold: Threshold::AbsoluteCount { weight: 2 },
+        required_weight: 2,
         max_voting_period: Duration::Height(3),
     };
