+++ contracts/cw3-flex-multisig/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cw3-flex-multisig"
-version = "0.11.0"
+version = "0.10.3"
 authors = ["Ethan Frey <ethanfrey@users.noreply.github.com>"]
 edition = "2018"
 description = "Implementing cw3 with multiple voting patterns and dynamic groups"
@@ -18,12 +18,11 @@ backtraces = ["cosmwasm-std/backtraces"]
 library = []
 
 [dependencies]
-utils = { path = "../../packages/utils", version = "0.11.0" }
-cw2 = { path = "../../packages/cw2", version = "0.11.0" }
-cw3 = { path = "../../packages/cw3", version = "0.11.0" }
-cw3-fixed-multisig = { path = "../cw3-fixed-multisig", version = "0.11.0", features = ["library"] }
-cw4 = { path = "../../packages/cw4", version = "0.11.0" }
-cw-storage-plus = { path = "../../packages/storage-plus", version = "0.11.0" }
+utils = { path = "../../packages/utils", version = "0.10.3" }
+cw2 = { path = "../../packages/cw2", version = "0.10.3" }
+cw3 = { path = "../../packages/cw3", version = "0.10.3" }
+cw4 = { path = "../../packages/cw4", version = "0.10.3" }
+cw-storage-plus = { path = "../../packages/storage-plus", version = "0.10.3" }
 cosmwasm-std = { version = "1.0.0-beta3" }
 schemars = "0.8.1"
 serde = { version = "1.0.103", default-features = false, features = ["derive"] }
@@ -31,5 +30,5 @@ thiserror = { version = "1.0.23" }
 
 [dev-dependencies]
 cosmwasm-schema = { version = "1.0.0-beta3" }
-cw4-group = { path = "../cw4-group", version = "0.11.0" }
-cw-multi-test = { path = "../../packages/multi-test", version = "0.11.0" }
+cw4-group = { path = "../cw4-group", version = "0.10.3" }
+cw-multi-test = { path = "../../packages/multi-test", version = "0.10.3" }
