+++ contracts/cw3-flex-multisig/src/error.rs
@@ -1,6 +1,4 @@
 use cosmwasm_std::StdError;
-use utils::ThresholdError;
-
 use thiserror::Error;
 
 #[derive(Error, Debug, PartialEq)]
@@ -8,8 +6,20 @@ pub enum ContractError {
     #[error("{0}")]
     Std(#[from] StdError),
 
-    #[error("{0}")]
-    Threshold(#[from] ThresholdError),
+    #[error("Invalid voting threshold percentage, must be in the 0.5-1.0 range")]
+    InvalidThreshold {},
+
+    #[error("Required quorum threshold cannot be zero")]
+    ZeroQuorumThreshold {},
+
+    #[error("Not possible to reach required quorum threshold")]
+    UnreachableQuorumThreshold {},
+
+    #[error("Required weight cannot be zero")]
+    ZeroWeight {},
+
+    #[error("Not possible to reach required (passing) weight")]
+    UnreachableWeight {},
 
     #[error("Group contract invalid address '{addr}'")]
     InvalidGroup { addr: String },
