+++ contracts/cw3-flex-multisig/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod contract;
-pub mod error;
+mod error;
 pub mod msg;
 pub mod state;
