+++ packages/cw3/src/lib.rs
@@ -6,6 +6,14 @@ mod query;
 pub use crate::helpers::Cw3Contract;
 pub use crate::msg::{Cw3ExecuteMsg, Vote};
 pub use crate::query::{
-    Cw3QueryMsg, ProposalListResponse, ProposalResponse, Status, VoteInfo, VoteListResponse,
-    VoteResponse, VoterDetail, VoterListResponse, VoterResponse,
+    Cw3QueryMsg, ProposalListResponse, ProposalResponse, Status, ThresholdResponse, VoteInfo,
+    VoteListResponse, VoteResponse, VoterDetail, VoterListResponse, VoterResponse,
 };
+
+#[cfg(test)]
+mod tests {
+    #[test]
+    fn it_works() {
+        // test me
+    }
+}
