+++ packages/multi-test/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "cw-multi-test"
-version = "0.11.0"
+version = "0.10.3"
 authors = ["Ethan Frey <ethanfrey@users.noreply.github.com>"]
 edition = "2018"
 description = "Test helpers for multi-contract interactions"
@@ -18,8 +18,8 @@ staking = ["cosmwasm-std/staking"]
 backtrace = ["anyhow/backtrace"]
 
 [dependencies]
-utils = { path = "../../packages/utils", version = "0.11.0" }
-cw-storage-plus = { path = "../../packages/storage-plus", version = "0.11.0"}
+utils = { path = "../../packages/utils", version = "0.10.3" }
+cw-storage-plus = { path = "../../packages/storage-plus", version = "0.10.3"}
 cosmwasm-std = { version = "1.0.0-beta3", features = ["staking"] }
 cosmwasm-storage = { version = "1.0.0-beta3" }
 itertools = "0.10.1"
