+++ packages/utils/src/lib.rs
@@ -4,7 +4,6 @@ mod expiration;
 mod pagination;
 mod parse_reply;
 mod payment;
-mod threshold;
 
 pub use pagination::{
     calc_range_end, calc_range_start, calc_range_start_string, maybe_addr, maybe_canonical,
@@ -15,7 +14,6 @@ pub use parse_reply::{
     ParseReplyError,
 };
 pub use payment::{may_pay, must_pay, nonpayable, one_coin, PaymentError};
-pub use threshold::{Threshold, ThresholdError, ThresholdResponse};
 
 pub use crate::balance::NativeBalance;
 pub use crate::event::Event;
